Huge changes - UNTESTED Adding Tree to get AI functionality off the ground Adding all copy constructors and Overloads for hashing so that the Tree dictionaries work

+++ Playable.cs
@@ -1,6 +1,6 @@
 namespace A1 {
     interface Playable {
         bool move (Board.direction dir);
-        void setCords(short setX, short setY);
+        void setCords(int setX, int setY);
     }
 }
